function/barcode/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-mgmt-communication
azure-communication-email
azure-communication-identity
azure-storage-blob
azure-data-tables
azure-identity
azure-keyvault-secrets
python-barcode
qrcode[pil]
Pillow
reportlab==4.0.7